+++ tablefill/__init__.py
@@ -5,12 +5,14 @@ __author__  = """Mauricio Caceres Bravo"""
 __email__   = 'mauricio.caceres.bravo@gmail.com'
 __version__ = '0.11.0'
 
-from .errors import InputParseError, PlaceholderError, TableFillError
+from .errors import DiagnosticContext, InputParseError, PlaceholderError, TableFillError, TemplateScanError
 from .tablefill import tablefill
 
 __all__ = [
+    'DiagnosticContext',
     'InputParseError',
     'PlaceholderError',
     'TableFillError',
+    'TemplateScanError',
     'tablefill',
 ]